requirements.txt
# ==========================================
# DEPENDENCIAS DEL PROYECTO: MEDICAL IOT
# ==========================================

# --- Streaming y Mensajería ---
kafka-python-ng==2.2.2

# --- Procesamiento de Datos y Matemáticas ---
pandas>=2.0.0
numpy>=1.24.0

# --- Inteligencia Artificial ---
scikit-learn>=1.3.0
mlflow>=2.10.0

# --- Base de Datos ---
# Nota: Se usa psycopg2-binary para evitar errores de compilación en Windows/Mac
psycopg2-binary>=2.9.9 

# --- Lectura de Datos Médicos ---
wfdb==4.1.2

# --- Serving Layer (API REST) ---
fastapi>=0.110.0
uvicorn>=0.28.0

# --- Configuración y Variables de Entorno ---
python-dotenv>=1.0.1